+++ src/core/shared/enums/Views.ts
@@ -82,3 +82,8 @@ export enum View_Events_Factions {
 export enum View_Events_Input_Menu {
     SetMenu = 'inputmenu:Set',
 }
+
+export enum View_Events_PaintShop {
+    Open = 'paint:Shop:Open',
+    Purchase = 'paint:Shop:Purchase',
+}