+++ src/core/shared/interfaces/IClientInteraction.ts
@@ -1,4 +1,5 @@
 export default interface IClientInteraction {
+    uid?: string;
     keyPress: string;
     description: string;
 }